+++ apps/io-wallet-user-func/src/attestation-service.ts
@@ -1,12 +1,9 @@
 import { FiscalCode, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import * as RTE from "fp-ts/ReaderTaskEither";
 import * as TE from "fp-ts/TaskEither";
-import { pipe } from "fp-ts/function";
 import { DeviceDetails } from "io-wallet-common/device-details";
 import { JwkPublicKey } from "io-wallet-common/jwk";
 
-import { AttestationServiceConfiguration } from "./app/config";
-import { MobileAttestationService } from "./infra/attestation-service";
 import { WalletAttestationRequest } from "./wallet-attestation-request";
 import { WalletInstanceRequest } from "./wallet-instance-request";
 
@@ -14,6 +11,10 @@ export enum OperatingSystem {
   android = "Android",
   iOS = "Apple iOS",
 }
+
+export type ValidationResult =
+  | { reason: string; success: false }
+  | { success: true };
 
 export interface ValidatedAttestation {
   deviceDetails: DeviceDetails;
@@ -31,6 +32,7 @@ export interface ValidateAssertionRequest {
 }
 
 export interface AttestationService {
+  getHardwarePublicTestKey: () => TE.TaskEither<Error, JwkPublicKey>;
   validateAssertion: (
     request: ValidateAssertionRequest,
   ) => TE.TaskEither<Error, void>;
@@ -45,21 +47,17 @@ export interface AttestationService {
 export const validateAttestation: (
   walletInstanceRequest: WalletInstanceRequest,
 ) => RTE.ReaderTaskEither<
-  { attestationServiceConfiguration: AttestationServiceConfiguration },
+  { attestationService: AttestationService },
   Error,
   ValidatedAttestation
 > =
   (walletInstanceRequest) =>
-  ({ attestationServiceConfiguration }) =>
-    pipe(
-      new MobileAttestationService(attestationServiceConfiguration),
-      (attestationService) =>
-        attestationService.validateAttestation(
-          walletInstanceRequest.keyAttestation,
-          walletInstanceRequest.challenge,
-          walletInstanceRequest.hardwareKeyTag,
-          walletInstanceRequest.fiscalCode,
-        ),
+  ({ attestationService }) =>
+    attestationService.validateAttestation(
+      walletInstanceRequest.keyAttestation,
+      walletInstanceRequest.challenge,
+      walletInstanceRequest.hardwareKeyTag,
+      walletInstanceRequest.fiscalCode,
     );
 
 export const validateAssertion: (
@@ -68,24 +66,18 @@ export const validateAssertion: (
   signCount: number,
   user: FiscalCode,
 ) => RTE.ReaderTaskEither<
-  { attestationServiceConfiguration: AttestationServiceConfiguration },
+  { attestationService: AttestationService },
   Error,
   void
 > =
   (walletAttestationRequest, hardwareKey, signCount, user) =>
-  ({ attestationServiceConfiguration }) =>
-    pipe(
-      new MobileAttestationService(attestationServiceConfiguration),
-      (attestationService) =>
-        attestationService.validateAssertion({
-          hardwareKey,
-          hardwareSignature:
-            walletAttestationRequest.payload.hardware_signature,
-          integrityAssertion:
-            walletAttestationRequest.payload.integrity_assertion,
-          jwk: walletAttestationRequest.payload.cnf.jwk,
-          nonce: walletAttestationRequest.payload.challenge,
-          signCount,
-          user,
-        }),
-    );
+  ({ attestationService }) =>
+    attestationService.validateAssertion({
+      hardwareKey,
+      hardwareSignature: walletAttestationRequest.payload.hardware_signature,
+      integrityAssertion: walletAttestationRequest.payload.integrity_assertion,
+      jwk: walletAttestationRequest.payload.cnf.jwk,
+      nonce: walletAttestationRequest.payload.challenge,
+      signCount,
+      user,
+    });