+++ apps/io-wallet-user-func/src/infra/attestation-service/errors.ts
@@ -1,24 +1,7 @@
-import { X509Certificate } from "crypto";
-import { AndroidDeviceDetails } from "io-wallet-common/device-details";
-
 export class AndroidAttestationError extends Error {
   name = "AndroidAttestationError";
-  constructor(
-    message: string,
-    {
-      deviceDetails,
-      x509Chain,
-    }: {
-      deviceDetails?: AndroidDeviceDetails;
-      x509Chain?: readonly X509Certificate[];
-    },
-  ) {
-    const finalMessage =
-      `[Android Attestation Error] ${message}` +
-      (x509Chain ? `\nx509Chain: ${JSON.stringify(x509Chain)}` : "") +
-      (deviceDetails
-        ? `\nDevice details: ${JSON.stringify(deviceDetails)}`
-        : "");
+  constructor(message: string) {
+    const finalMessage = `[Android Attestation Error] ${message}`;
 
     super(finalMessage);
   }