+++ infra/resources/_modules/storage/storage_queue.tf
@@ -1,0 +1,4 @@
+resource "azurerm_storage_queue" "wallet_instances_revocation_check" {
+  name                 = "wallet-instances-revocation-check"
+  storage_account_name = azurerm_storage_account.wallet_revocation_storage.name
+}